Moved SaveGameTests to owm package, fixed tests
